# api/openapi/appointment-service.openapi.yaml
openapi: 3.0.3
info:
  title: Appointment Service API
  version: 1.0.0
  description: |
    API for creating and listing medical appointments. The service assigns an available
    doctor (by specialty) and an available room automatically, and triggers post-booking actions.
servers:
  - url: http://localhost:8080
tags:
  - name: Appointments
paths:
  /appointments:
    post:
      tags: [Appointments]
      summary: Create an appointment (doctor + room assigned automatically)
      operationId: createAppointment
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: |
            Optional idempotency key to safely retry appointment creation without duplicating bookings.
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentRequest'
            examples:
              example:
                value:
                  patientName: "Jane Doe"
                  patientEmail: "jane.doe@example.com"
                  specialty: "DERMATOLOGY"
                  startTime: "2026-02-11T10:00:00Z"
                  endTime: "2026-02-11T10:30:00Z"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: No availability (doctor or room) for the requested timeslot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: [Appointments]
      summary: List scheduled appointments (admin)
      operationId: listAppointments
      parameters:
        - name: from
          in: query
          required: false
          description: Filter appointments with endTime >= from
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          description: Filter appointments with startTime <= to
          schema:
            type: string
            format: date-time
        - name: specialty
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/AppointmentStatus'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: A paged list of appointments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentListResponse'
        "400":
          description: Invalid query params
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /appointments/{appointmentId}:
    get:
      tags: [Appointments]
      summary: Get appointment details by id
      operationId: getAppointment
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Appointment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    CreateAppointmentRequest:
      type: object
      required: [patientName, patientEmail, specialty, startTime, endTime]
      properties:
        patientName:
          type: string
          minLength: 1
        patientEmail:
          type: string
          format: email
        specialty:
          type: string
          minLength: 1
          description: Medical specialty required for the appointment.
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time

    AppointmentResponse:
      type: object
      required: [id, doctor, room, patientName, patientEmail, specialty, startTime, endTime, status]
      properties:
        id:
          type: string
          format: uuid
        doctor:
          $ref: '#/components/schemas/AssignedDoctor'
        room:
          $ref: '#/components/schemas/AssignedRoom'
        patientName:
          type: string
        patientEmail:
          type: string
          format: email
        specialty:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/AppointmentStatus'

    AssignedDoctor:
      type: object
      required: [externalId, name, specialty]
      properties:
        externalId:
          type: string
          description: Optional reference to the external doctor system.
        name:
          type: string
        specialty:
          type: string

    AssignedRoom:
      type: object
      required: [externalId, name]
      properties:
        externalId:
          type: string
          description: Optional reference to the external room system.
        name:
          type: string

    AppointmentStatus:
      type: string
      description: Current appointment state.
      enum: [PENDING, CONFIRMED, FAILED, CANCELLED]

    AppointmentListResponse:
      type: object
      required: [items, limit, offset, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AppointmentResponse'
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
          description: Total matching appointments (for pagination).

    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable error code for clients (e.g., VALIDATION_ERROR, NO_AVAILABILITY).
        message:
          type: string
        details:
          type: object
          additionalProperties: true